Online order booking
This app is basically a project of an art gallery. Where user can book order online and send there code or order no and complete details on whatsapp and SMS. Also user can maintain a record how many frames are done or how many frames are delivered. A user can check the data or daily income on mobile phone